extract to utility
