+++ src/Routes.js
@@ -1,4 +1,4 @@
-import {includes, mergeAll} from 'lodash/fp';
+import {mergeAll} from 'lodash/fp';
 import React from 'react';
 import {Route, Switch} from 'react-router-dom';
 import AuthenticatedRoute from './components/AuthenticatedRoute';
@@ -40,34 +40,7 @@ import DataSubmissionForm from './pages/DataSubmissionForm';
 import {ensureSoHasDaaAcknowledgement} from './components/SigningOfficialDaaAgreementWrapper';
 import BroadDatasetCatalog from './pages/dac_dataset_catalog/BroadDatasetCatalog';
 import {AnVILDMSPolicyInfo, NIHDMSPolicyInfo} from './pages/DMSPolicyInfo';
-import {Storage} from './libs/storage';
-
-/**
- * Predefined groups of environments for which certain features are valid for.
- * @type {{NON_STAGING: string[], NON_PROD: string[]}}
- */
-export const envGroups = {
-  NON_PROD: ['local', 'dev', 'alpha', 'staging'],
-  NON_STAGING: ['local', 'dev', 'alpha']
-};
-
-/**
- * Returns true if the current application `Storage.ENV` variable exists as an element
- * in the specified environment group or false otherwise. Used to determine if a route
- * should be displayed or not based on current env.
- *
- * @example
- * // returns true when Storage.ENV === 'dev' | 'local'
- * checkEnv(envGroups.NON_STAGING)
- * // returns false when Storage.ENV === 'staging' | 'prod'
- * checkEnv(envGroups.NON_STAGING)
- * @param envGroup
- * @returns {boolean}
- */
-export const checkEnv = (envGroup) => {
-  const env = Storage.getEnv();
-  return env ? includes(env)(envGroup) : false;
-};
+import {checkEnv, envGroups} from './utils/RouteUtils';
 
 const Routes = (props) => (
   <Switch>
